fix:🔥 lat lng is float

+++ movements/models.py
@@ -5,8 +5,8 @@ from patients.models import Patient
 class Movement(models.Model):
     date = models.DateField()
     place = models.CharField(max_length=20)
-    lat = models.IntegerField()
-    lng = models.IntegerField()
+    lat = models.FloatField()
+    lng = models.FloatField()
     patient = models.ForeignKey(
         Patient, related_name="movements", on_delete=models.CASCADE
     )
